requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pytest
typing_extensions
anyio
starlette
azure-identity
azure-keyvault-secrets
azure-storage-blob